+++ src/sagemaker/model.py
@@ -303,13 +303,14 @@ class Model(object):
                            volume_kms_key=volume_kms_key, sagemaker_session=self.sagemaker_session)
 
     def delete_model(self):
-        """Delete an Amazon SageMaker ``Model``.
+        """Delete an Amazon SageMaker Model.
 
-        Raises: ValueError if model is not deployed yet.
+        Raises:
+            ValueError: if the model is not created yet.
 
         """
         if self.name is None:
-            raise ValueError('The SageMaker model must be deployed first before attempting to delete.')
+            raise ValueError('The SageMaker model must be created first before attempting to delete.')
         self.sagemaker_session.delete_model(self.name)