+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 CHANGELOG
 =========
 
+1.18.4.dev
+==========
+
+* doc-fix: Remove incorrect parameter for EI TFS Python README
+
 1.18.3.post1
 ============